+++ Sources/StreamChat/WebSocketClient/Events/EventType.swift
@@ -42,6 +42,8 @@ public extension EventType {
     // MARK: Channel Events
     
     /// When a channel was updated.
+    static let channelCreated: Self = "channel.created"
+    /// When a channel was updated.
     static let channelUpdated: Self = "channel.updated"
     /// When a channel was deleted.
     static let channelDeleted: Self = "channel.deleted"
@@ -122,6 +124,7 @@ extension EventType {
         case .userUnbanned:
             return try (try? UserUnbannedEventDTO(from: response)) ?? UserGloballyUnbannedEventDTO(from: response)
 
+        case .channelCreated: throw ClientError.IgnoredEventType()
         case .channelUpdated: return try ChannelUpdatedEventDTO(from: response)
         case .channelDeleted: return try ChannelDeletedEventDTO(from: response)
         case .channelHidden: return try ChannelHiddenEventDTO(from: response)